+++ Source/DafnyCore/Resolver/GhostInterestVisitor.cs
@@ -523,8 +523,6 @@ class GhostInterestVisitor {
           break;
         }
       case BlockByProofStmt blockByProofStmt:
-        // TODO move to the BlockByProofStmt class
-
         blockByProofStmt.IsGhost = mustBeErasable;  // set .IsGhost before descending into substatements (since substatements may do a 'break' out of this block)
         Visit(blockByProofStmt.Body, mustBeErasable, proofContext);
         blockByProofStmt.IsGhost = blockByProofStmt.IsGhost || blockByProofStmt.Body.IsGhost;
