+++ Source/DafnyCore/Resolver/TailRecursion.cs
@@ -268,8 +268,6 @@ class TailRecursion {
     } else if (stmt is VarDeclPattern) {
     } else if (stmt is ExpectStmt) {
     } else if (stmt is BlockByProofStmt blockByProofStmt) {
-      // TODO something with the by block?
-      // TODO Move this to BlockByProofStmt class ?
       return CheckTailRecursive(blockByProofStmt.Body, enclosingMethod, ref tailCall, reportErrors);
     } else {
       Contract.Assert(false);  // unexpected statement type
