+++ Source/DafnyCore/Verifier/Variables.cs
@@ -1,0 +1,9 @@
+using Microsoft.Boogie;
+
+namespace Microsoft.Dafny;
+
+public class Variables : OrderedDictionary<string, Variable> {
+  public Variables() : base(v => v.Name) {
+  }
+
+}
